# docs/identity/openapi.yaml
openapi: 3.1.0
info:
  title: Platform Demo Identity API
  version: 1.7.0
  summary: Current identity API plus planned auth and federation contract.
  description: |
    Contract-first definition for the identity module.

    Implemented now:
    - Register user
    - Get user by id
    - Login with local credentials
    - Refresh token rotation
    - Logout session revocation
    - Session read endpoint via bearer access token

    Planned next:
    - harden auth telemetry/audit events
    - OIDC auth endpoints
    - provider connection management endpoints
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /v1/refresh:
    post:
      tags: [Auth]
      summary: Rotate refresh token and mint a new access token
      description: |
        Validates refresh token state, rotates refresh token, and mints a new access token.
        Refresh token reuse after rotation revokes the session token family.
      operationId: refresh
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Token pair rotated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenSessionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/logout:
    post:
      tags: [Auth]
      summary: Revoke the current refresh token/session
      description: |
        Revokes the session tied to the provided refresh token and revokes
        all refresh tokens in that session family.
      operationId: logout
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogoutRequest'
      responses:
        '204':
          description: Logged out and refresh token revoked
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/session:
    get:
      tags: [Auth]
      summary: Get the current authenticated session context
      description: |
        Requires a bearer access token validated by resource-server rules.
        Intended for `@PreAuthorize`-based method-level authorization flows with a session-oriented response.
      operationId: getSession
      responses:
        '200':
          description: Current session context
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
...
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
    LogoutRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
...
    TokenSessionResponse:
      type: object
      required: [access_token, token_type, expires_in, refresh_token, refresh_expires_in]
      properties:
        access_token:
          type: string
        token_type:
          type: string
          const: Bearer
        expires_in:
          type: integer
          example: 900
        refresh_token:
          type: string
        refresh_expires_in:
          type: integer
          example: 2592000
        scope:
          type: string
    SessionResponse:
      type: object
      required: [session_id, user, provider, session_status, access_token_expires_in]
      properties:
        session_id:
          type: string
        user:
          $ref: '#/components/schemas/User'
        provider:
          type: string
          example: LOCAL
        session_status:
          type: string
          enum: [ACTIVE, EXPIRED, REVOKED]
        access_token_expires_in:
          type: integer
        refresh_token_expires_in:
          type: [integer, 'null']
...
    User:
      type: object
      required: [id, email, status, email_verified, created_at]
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        status:
          type: string
          enum: [ACTIVE, INACTIVE, DEACTIVATED]
        email_verified:
          type: boolean
        created_at:
          type: string
          format: date-time
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code:
              type: string
            message:
              type: string
            request_id:
              type: string
            details:
              type: object
              additionalProperties: true